Switch to merge-anything for merging contexts
This fixes #5 and we can now log URL objects.

To make this work I had to update all the things, which means I
need to add an explicit cast to the various useXXX functions.
I'd like to come back and fix that properly with some TS magic
but life is short, and it's a Saturday.